Add a Line array and handle array resize on both Line and Pipe
Added a pipes for a total of 6.
Line array is going to be replaced by a Map.  Doing this because
children cannot be added in an enclosed scope.
"Local variable root defined in an enclosing scope must be final or
effectively final"
Also determine that only 1 custom part per java file.  However you can
put multiple java files in a .jar.
Also note that some lines aren't being shown.