Trig functions to set width and height
The calculations for width and height are right.  The new pipe is is
intended to be an angled one from right/top to left/bottom but displays
as a horizontal pipe. We need to play around with setting rectangle
values such as x, y, layoutX and LayoutY and likely rotate.