Add listeners for all head and tail Lines.
Minimize still fails.  There were 24 calls to processAnchors() - 6 pipes
each with a head and tail (12) then called on both layoutXProperty() and
layoutYProperty() (24)

TODO pass the Line() to processAnchors so it can update only pipes that
have that head or tail.

It looks like we may have to get Oracle support.